% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensembl_db_retrieval.R
\name{ensembl_db_retrieval}
\alias{ensembl_db_retrieval}
\title{Retrieve Transcript Information from Ensembl for all Primary Transcripts}
\usage{
ensembl_db_retrieval(organism)
}
\arguments{
\item{organism}{String indicating if mouse or human transcripts should be
retrieved}
}
\value{
A data frame containing the gene names, transcript IDs, APPRIS
annotations, UniProt Swissprot IDs, UniProt Swissprot isoform IDs, and
UniProt TREMBL IDs for all the primary transcripts in an organism.
}
\description{
This function retrieves all the primary transcripts in the given organism
and their corresponding gene names, APPRIS annotations, and UniProt IDs.
}
