% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_ids.R
\name{format_ids}
\alias{format_ids}
\title{Reformat transcripts to facilitate fasta file conversion}
\usage{
format_ids(final_pairs)
}
\arguments{
\item{final_pairs}{Data frame containing original row-wise pairings of
primary and alternative transcripts for inputted data without associated
sequences}
}
\value{
A data frame containing the gene names, transcript IDs, APPRIS
annotations, UniProt Swissprot IDs, UniProt Swissprot isoform IDs, and
UniProt TREMBL IDs for all the given and associated primary transcripts in
an alternating fashion
}
\description{
Modify format of data to display all primary and alternative transcripts
from the same gene together and remove any duplicates.
}
