% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPA_data.R
\docType{data}
\name{hpa_genes}
\alias{hpa_genes}
\title{SurfaltR Gene Name and Transcript ID Test Data- Highly Expressed Human 
Alternative Transcripts}
\format{
A data frame with 10 rows and 2 variables:
\describe{
\item{gene_name}{Gene name corresponding to transcript ID}
\item{transcript}{transcript ID of gene of interest}
...
}
}
\source{
\url{https://science.sciencemag.org/content/347/6220/1260419/tab-figures-data}
}
\usage{
hpa_genes
}
\description{
For the gene name and transcript ID input, we have included 10 unique human
transcripts from 7 different genes annotated as alternative by APPRIS. These
genes were derived from supplementary data 12 from Uhlén et all 2015.
This data has been formatted to be compatible with the package.
}
\keyword{datasets}
