% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPA_mouse_data.R
\docType{data}
\name{hpa_mouse_genes}
\alias{hpa_mouse_genes}
\title{SurfaltR Gene Name and Transcript ID Test Data- Highly Expressed Mouse Alternative Transcripts}
\format{
A data frame with 5 rows and 2 variables:
\describe{
  \item{gene_name}{Gene name corresponding to transcript ID}
  \item{transcript}{transcript ID of gene of interest}
  ...
}
}
\source{
\url{https://science.sciencemag.org/content/347/6220/1260419/tab-figures-data}
}
\usage{
hpa_mouse_genes
}
\description{
For the gene name and transcript ID input, we have included 5 unique mouse 
transcripts from 5 different genes annotated as alternative by APPRIS. These 
genes were derived from supplementary data 12 from Uhlén et all 2015. This 
data has been formatted to be compatible with the package and to match the 
genes in the HPA human gene dataset.
}
\keyword{datasets}
