% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_surfaltr.R
\name{test_surfaltr}
\alias{test_surfaltr}
\title{Test the functionality of SurfaltR}
\usage{
test_surfaltr()
}
\value{
Nothing is returned.
}
\description{
This function runs all of SurfaltR's other functions on the CRB1 data set
to ensure that the function output matches the expected output. An incorrect
output or error indicates that something went wrong in installation.
}
\note{
If the results from the test match the expected
results, a message stating that the test worked will be printed. If not, the
user will be prompted to check the installation
}
\examples{
\donttest{
test_surfaltr()
}
}
