% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmhmm_fix_path.R
\name{tmhmm_fix_path}
\alias{tmhmm_fix_path}
\title{Retrieve Data from TMHMM and Fix Functionality of TMHMM R Package}
\usage{
tmhmm_fix_path(fasta_filename, folder_name = tmhmm::get_default_tmhmm_folder())
}
\arguments{
\item{fasta_filename}{Parameter containing input fasta file to be run on
tmhmm}

\item{folder_name}{Path to folder containing installed tmhmm software}
}
\value{
Raw results from tmhmm containing membrane locations for
each transcript
}
\description{
This function retrieves the raw data from tmhmm containing information
about the membrane location of each amino acid in a transcript.
In order to set a standard path that allows tmhmm to run, the path
is set to match that of the fasta file contining the amino acids.
}
\note{
In order for this function to work, there needs to be a .fasta file
containing the amino acid sequences for each transcript called "AA.fasta"
saved to a folder called output within the working directory.
}
